app refactoring -- component reorganization and state lifting into container components
